backend/requirements-dev.txt
# ==============================================================================
# EUNOMIA Legal AI Platform - Development Dependencies (Windows)
# ==============================================================================
# Python: 3.13+ compatible
# Usage: pip install -r requirements-dev.txt

# Core Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.17

# Database - PostgreSQL Async (NO psycopg2-binary for Windows)
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
# psycopg2-binary excluded - requires PostgreSQL libs on Windows

# Database - Redis
redis==5.2.0
# hiredis excluded - requires C compiler on Windows

# Security & Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# Celery - Task Queue
celery==5.4.0
flower==2.0.1

# Vector Database - Qdrant
qdrant-client==1.12.1

# AI/ML - Hugging Face Transformers (CPU only for Windows)
transformers==4.46.3
torch==2.7.1  # Latest version compatible with Python 3.13
sentencepiece==0.2.0
protobuf==5.29.1
accelerate==1.2.1

# AI/ML - Sentence Transformers (Embeddings)
sentence-transformers==3.3.1

# AI/ML - Ollama Client
ollama==0.4.4

# Document Processing
pypdf==5.1.0
python-docx==1.1.2
filetype==1.2.0  # Alternative to python-magic (pure Python)

# HTTP Client
httpx==0.28.1
aiohttp==3.11.10

# Validation & Serialization
pydantic-core==2.23.4
email-validator==2.2.0

# Logging & Monitoring
loguru==0.7.3
prometheus-client==0.21.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
faker==33.1.0

# Development Tools
black==24.10.0
ruff==0.8.4
mypy==1.13.0
pre-commit==4.0.1

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0
pydantic-extra-types==2.10.1

# Date & Time
python-dateutil==2.9.0.post0
pytz==2024.2

# File Storage
aiofiles==24.1.0

# NLP - Spacy (Optional in dev)
spacy==3.8.2

# ==============================================================================
# DEVELOPMENT-ONLY PACKAGES
# ==============================================================================
ipython==8.31.0
ipdb==0.13.13
rich==13.9.4
watchfiles==1.0.0

# ==============================================================================
# NOTES FOR WINDOWS DEVELOPMENT
# ==============================================================================
# 1. PostgreSQL: Use Docker
#    docker run -d -p 5432:5432 -e POSTGRES_PASSWORD=dev postgres:16-alpine
#
# 2. Redis: Use Docker
#    docker run -d -p 6379:6379 redis:7-alpine
#
# 3. Qdrant: Use Docker
#    docker run -d -p 6333:6333 qdrant/qdrant
#
# 4. Ollama: Download from https://ollama.com for Windows